+++ gameshark/thigsgen.c
@@ -1,7 +1,6 @@
 #include <stdio.h>
 #include <stdlib.h>
 #include <string.h>
-#include <stdbool.h>
 #include "thigsgen.h"
 
 FILE *out;
@@ -16,9 +15,6 @@ unsigned int gameshark_code_address;
 unsigned int gameshark_code_address_2;
 unsigned int gameshark_code_address_mod_val;
 unsigned int gameshark_code_address_mod_comparison_val;
-
-bool mcs = 0;
-bool raw = 0;
 
 void parse() {
     printf("GS Code Addr: %08X\n", gameshark_code_address);
@@ -66,60 +62,39 @@ void parse() {
 int main (int argc, const char * argv[]) 
 {
 
-    printf("Tonyhax International GameShark Generator (THIGSGEN) v1.0.1 By Alex Free\n");
+    printf("Tonyhax International GameShark Generator (THIGSGEN) v1.0.2 By Alex Free\n");
 
-   if(argc != 4)
+   if(argc != 2)
     {
-        printf("Error, THIGSGEN requires 3 arguments.\nUsage:\n\nthigsgen <output mode> <txt file containing GameShark codes> <output save file>\n\n<output mode> -m for .mcs format. -r for raw format.\n\n<output save file> must be TONYHAXINTGS for raw format. It can however be anything you want for mcs format\n\nExamples:\n\nthigsgen -r gs.txt TONYHAXINTGS\n\nthigsgen -m gs.txt tonyhax-codes.mcs\n");
+        printf("Error, THIGSGEN requires 1 argument.\nUsage:\n\nthigsgen <txt file containing GameShark codes>\n");
         return(1);
     }
 
     
-    if((strcmp("-m", argv[1])) == 0) 
-    {
-      mcs = 1;
-      printf("Output format: MCS\n\n");
-    } else if((strcmp("-r", argv[1])) == 0) {
-      raw = 1;
-      printf("Output format: RAW\n\n");
-    } else {
-      printf("Error: 1st argument must be -r or -m\n");
-      return(1);
-    }
-    
-    if((txt = fopen(argv[2], "r")) == NULL)
+    if((txt = fopen(argv[1], "r")) == NULL)
     {
         printf("Error: Cannot read the input txt file: %s\n", argv[2]);
         return(1);
     }
 
-    if((out = fopen(argv[3], "wb+")) == NULL)
+    if((out = fopen("TONYHAXINTGS", "wb+")) == NULL)
     {
-        printf("Error: Cannot create the output file: %s\n", argv[3]);
+        printf("Error: Cannot create the output save file TONYHAXINTGS\n");
         fclose(txt);
         return(1);
     }
 
-
-
     fseek(out, 0, SEEK_SET);
     fseek(txt, 0, SEEK_SET);
 
-    if(mcs)
-    {
-      for(int i = 0; i < 128; i++)
-        fputc(mcshead[i], out); 
-    }
+    for(int i = 0; i < 0x2000; i++)
+      fputc(0x00, out); 
 
-    for(int i = 0; i < 8192; i++)
+    fseek(out, 0, SEEK_SET);
+    for(int i = 0; i < 0x101; i++)
       fputc(save[i], out); 
     
-    if(mcs)
-    {
-      fseek(out, 0x181, SEEK_SET);
-    } else if(raw) {
-      fseek(out, 0x101, SEEK_SET);
-    }
+    fseek(out, 0x103, SEEK_SET);
 
 		while(1) 
     {
@@ -135,16 +110,11 @@ int main (int argc, const char * argv[])
 				}
     }	
 
-    if(mcs)
-    {
-      fseek(out, 0x180, SEEK_SET);
-    } else if(raw) {
-      fseek(out, 0x100, SEEK_SET);
-    }
+    fseek(out, 0x102, SEEK_SET);
 
     fwrite(&number_of_gameshark_codes, 1, 1, out);
     fclose(out);
     fclose(txt);
-    printf("Generated Tonyhax International save file named: %s containing %d GameShark Codes\n", argv[3], number_of_gameshark_codes);
+    printf("Generated Tonyhax International save file named: TONYHAXINTGS containing %d GameShark Codes\n", number_of_gameshark_codes);
 
 }